[PORT FROM R41][WORKAROUND][AUDIO] Adding vibrator duration reduction logic
BZ: 71799

Depending on the hardware type, the vibration strengh can be different.
And it could make some noise to remote party when the phone call is
activated and user types something in message application.

This patch will force to reduce the vibration duration with android
property(Audiocomms.Vibrator.ReduceRate).
If the property is not defined, it will not reduce the duration.

Change-Id: I55c3024d5ede3a7310cdc89d9c5cf7d4b048f976
Origin-Change-Id: If5a248456c3353f13e1db7e23b0a1f616477628b
Signed-off-by: Jhin Lee <jhinx.lee@intel.com>
Reviewed-on: http://android.intel.com:8080/84300
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>