[AUDIO] new way to check if a device is used as input
BZ: 71815

There is no way to know if a device is an input or an output.
Check patch 79263 : the direction bit will be removed from the
device value.
So an other handle argument should be used to check if the device
is an input or on output.
curFmRxMode is set as UNSET in default constructors _defaultsIn and
_defaultsOut. This is needed to initialize isOutput.

Change-Id: I20491e9c1cb8b7c62faabf4d4dd6e104ddb10f37
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/78890
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>