Add compositionComplete interface to hack FBO flush issue
BZ: 72011

PVR's OGL driver will not flush FBO when switch from FBO to normal
framebuffer. Before JB MR1, SurfaceFlinger will call compositionComplete
to flush FBO in captrueScreen function. But in JB MR1, FB HAL is closed
and SurfaceFlinger will not call compositionComplete to flush FBO. This
is a temp hack to let FBO flush in this case. The patch should be revert
if IMG provide a formal fixing for this issue.

Change-Id: I443a71216379329676cb810c5cd3572658c5ecb8
Signed-off-by: Marc Mao <marc.mao@intel.com>
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/79389
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>