HWC: fix scroll bar displayed on HDMI in video ext mode
BZ: 74798

This is the patch to fix scroll bar displayed on HDMI in video Ext
mode. It is due to in mfld sprite plane flip function, previously
it flips all connected planes instead of its own index plane. As
in MR1, hdmi flip has been moved into hwc, this needs to be corrected.

Signed-off-by: Lei Zhang <lei.zhang@intel.com>

Change-Id: I8c11fdb2bb37ec9356a824602f3ea8498358cc7d
Reviewed-on: http://android.intel.com:8080/83159
Reviewed-by: Hu, Austin <austin.hu@intel.com>
Reviewed-by: Zhang, Lei <lei.zhang@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>