HWC: Fix the bug that Video isn't fluent when rotating DUT.
BZ: 81271

It is a regression caused by the patch 86000.
The patch 86000 is for fixing overlay crash and grenn screen,
But I couldn't find the root cause of green screen, after discuss with
GFX team, I used a workaround to fix it, and this workaround caused this
regression.
Now, we had found there are 2 reasons of green screen:
1:When Surfaceflinger release HWC, HWC should release its
layers, disbale all pipes and planes;
2:SF still calls "prepare" and "commit" APIs to flip data although HWC
is released, It may be SF's bug;

Change-Id: I82b87ab6f3b1e80bf0aad6240cc6ef91cf174d2a
Signed-off-by: Zhu,Tianyang <tianyang.zhu@intel.com>
Reviewed-on: http://android.intel.com:8080/87198
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>