[Audio] Fix ad_proxy to handle new AuVid protocol for es325
BZ: 83483

With Audience es325, Audience introduced a new version of the computer
base tuning tool AuVid.

The new version of AuVid does not communicate the same way than
previous versions. It does not send packets anymore, but directly
bytes to be sent to Audience chip es325. That means that for each
bytes received, the proxy now has to sent them to the chip, and
read the chip response which consist in reading as many bytes as
written bytes.

This is an enabler patch to get the AuViD chain working with es325.
A long term solution is now under investigation to redesign the
tool chain for Audience.

The patch updates the proxy to detect the Audience chip in use
(es305b or es325) thanks to the property audiocomms.vp.fw_name.
In case of es325, the packet management is not used and the bytes
received from AuViD are directly sent to the chip. Then, each time
bytes are written, the same amount of bytes is read from the chip
and sent to AuViD as response.

The ad_proxy user guide remains the same, and the procedure to use
AuViD and proxy remains the same than with the es305.

Change-Id: I808855525c14261fc4862a863538c6587137fb2e
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/90320
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lee, JhinX <jhinx.lee@intel.com>
Reviewed-by: Bellouard, Jean-marie <jean-marie.bellouard@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>