Widi: Remove dependency of Widi Extended mode switch on mHotPlugEvent
BZ: 84704

This patch makes the check for player status, to switch to WiDi Extended
video mode, independent of mHotPlugEvent. mHotPlugEvent is anyway meant
only for HDMI and Widi should not be dependent on it. This dependency also
causes problems like bug 84704 where there is a delta between mHotPlugEvent
being set to true and Widi becoming active.

Change-Id: I7f5a953e1b9f4a782291183f7a5d82783e27eb55
Signed-off-by: Rajneesh Chowdhury <rajneesh.chowdhury@intel.com>
Reviewed-on: http://android.intel.com:8080/90960
Reviewed-by: Singhi, Ashish <ashish.singhi@intel.com>
Reviewed-by: Veeramani, Karthik <karthik.veeramani@intel.com>
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Reviewed-by: Zhang, Lei <lei.zhang@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>