widi: Implement HWC 1.2 and rewrite extended video mode
BZ: 86947

The HWC 1.2 interface adds the ability to provide a 'Virtual Display'
to SurfaceFlinger in addition to the primary and external displays
supported by the HWC 1.1 interface. This new display type will receive
prepare and commit calls when the Wireless Display feature is in use.

By implementing HWC 1.2 and moving the extended video mode feature to
the virtual display, we gain the following benefits:

1. Video playback meant only for the MIPI won't trigger extended video
   mode so it won't override a Presentation that is supposed to be
   controlling the content on the remote display.

2. While extended mode is active, since we'll be using the video
   frames directly and not the framebuffer, we can tell SurfaceFlinger
   not to composite for the wireless display by setting all the layers
   to HWC_OVERLAY.

3. It will be possible now to use extended mode with a Presentation's
   content. However, we don't do this currently in case there is other
   content besides the video in the Presentation.

With this change, we can re-do video multitasking as an app that plays
a video in a Presentation, without having the problem that video
playback on the MIPI would activate extended mode and override the
background video player.

Extended video mode is now implemented almost entirely within the new
WidiDisplayDevice class, which also impelements the IFrameServer
interface, and the obsolete extended video mode code is removed.

The only widi code that still exists in IntelMIPIDisplayDevice just
handles the mipi-related aspects of extended mode (mipi turnoff and
suppressing video display). Also in HWComposer, prepareDisplays()
has been modified to check for a video layer that is displayed on
both mipi and widi, in order to activate extended mode. It also
calls prepare() on the widi device first, which allows widi to
prevent extended mode from being used.

Change-Id: Ib7adcffc2d1fbfefdc8ef2a1d7f0862b71da4b9e
Signed-off-by: Brian Rogers <brian.e.rogers@intel.com>
Reviewed-on: http://android.intel.com:8080/91800
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Mao, Marc <marc.mao@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>