widi: Fixed the binder bug to return correct error type.
BZ: 91319
When playing a video in extended mode bufferAvailable
sometimes returns error code which was not correctly
mapped, it was returning NO_ERROR for all case, as a
result buffer will be considered as being held by widi.

As a result of this decoder will run out of the buffer
bot "can't play this video" will popped up.

This has been fixed by returing ther correct error type.
Signed-off-by: Ashish Singhi <ashish.singhi@intel.com>

Change-Id: I7ca80b2f3f4b088086cc2b510d2a3d892445f9f3
Reviewed-on: http://android.intel.com:8080/96937
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Veeramani, Karthik <karthik.veeramani@intel.com>
Reviewed-by: Singhi, Ashish <ashish.singhi@intel.com>
Reviewed-by: Chowdhury, Rajneesh <rajneesh.chowdhury@intel.com>
Reviewed-by: Rogers, Brian E <brian.e.rogers@intel.com>
Reviewed-by: Mao, Marc <marc.mao@intel.com>
Reviewed-by: Zhang, Lei <lei.zhang@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>