[Audio] Update Audience chip detection in ad_streamer
BZ: 89365

Before to implement bug 89365, this patch cleans the way
ad_streamer detects the Audience chip.

Previously, it was detected checking a property supposed to
contain the chip version.

It now uses a chip identification command supported by all
Audience chips. The chip replies to this command with a unique
ID specific to each Audience chip family, allowing to detect
if the chip is supported by ad_streamer.

Change-Id: Ie946ebbd4d1d948dc25c1a70303c1544bedbed57
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/97087
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Quervel, Thierry <thierry.quervel@intel.com>
Reviewed-by: Continente, Christophe <christophe.continente@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>