[Audio] Enable dual channel capture on es325
BZ: 89365

This patch enables the dual channel support of ad_streamer on
Audiece chip es325.

On es305, one or two channel IDs have to be or'ed in a bit field
in order to select two channels capture.

On es325, two subsequent channel set commands have to be sent
to the Audience chip, one for each channel, before to start the
streaming.

This patch updates the way the user provides channel IDs from
the command line and update ad_streamer.c to support one or
two channel(s) capture on both es305 & es325 chips.

Change-Id: I000268008c70a0fe33f1e220b27f99edbf7e43be
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/97110
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Quervel, Thierry <thierry.quervel@intel.com>
Reviewed-by: Continente, Christophe <christophe.continente@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>