[PORT FROM MAIN] Fix wrong smart composition condition for video playback
BZ: 96412

For video playback case, video layer's compositionType is not always
set to HWC_OVERLAY in geometry change time. So mHasGlesCompostion is
wrong when in count video layers. But in actually, we only want to
enter smart composition when YUV is on overlay and has RGB to be
composited to framebuffer.

Change-Id: I44e8feebbd7d8648f4274004beb2f0d3dbba0388
Signed-off-by: Marc Mao <marc.mao@intel.com>
Reviewed-on: http://android.intel.com:8080/100417
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>