[CTP-All] Klocwork critical in HDMI-Audio HAL
BZ: 97058

Unvalidated integer value '*readbuffer' is received from an external
function through a call to 'fread'.In this patch strtol() function is
added instead of atoi() to convert the data in readbuffer to integer.
This function will handle the error condition properly if the 'readbuffer'
contains any invalid data.Added NULl check for some variables to avoid
NULL pointer dereference.

Change-Id: I8aa8c14821f2a2801dcbf43702e075e57fb3edee
Signed-off-by: sandhya <sandhyax.sebastian@intel.com>
Reviewed-on: http://android.intel.com:8080/99285
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Reviewed-by: Kalugin, Evgeny <evgeny.kalugin@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>