[PORT FROM MAIN][CTP-All] Klocwork critical in HDMI-Audio HAL
BZ: 97058

Unvalidated integer value '*readbuffer' is received from an external
function through a call to 'fread'.In this patch strtol() function is
added instead of atoi() to convert the data in readbuffer to integer.
This function will handle the error condition properly if the 'readbuffer'
contains any invalid data.Added NULl check for some variables to avoid
NULL pointer dereference.

Change-Id: If2f73c471bcfe556a3a371fd569f4cd2377f8e87
Orig-Change-Id: I8aa8c14821f2a2801dcbf43702e075e57fb3edee
Signed-off-by: sandhya <sandhyax.sebastian@intel.com>