[PORT FROM MAIN][AD_PROXY] Isolate es325 protocol handling
BZ: 89322

The new AuViD protocol introduced with es325 is different than the one
which was previously handled for es305b.

The enabling of es325 has been done in the same function of the es305b.

In order to ease maintenability, the protocol (old and new) are each in
a specific function.

Change-Id: I1e32eb93e4bd4c8d9862b898f3b5f0a90b745aca
Origin-Change-Id: Icaa87cb613cbc8c4837a4cebfa24e38db9ac57aa
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/109046
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>