[PORT FROM MAIN][AD_PROXY] No more support old packet protocol
BZ: 89322

The latest AuViD version for es305b does not use the packet protocol
anymore, but the new protocol already in-use with es325.

This patch removes the handling of the old packet protocol.

In consequence, the user shall updates its AuViD software for es305b
in order to use the proxy. The user using an AuViD for es325 doens't
have any update to do: its AuViD for es325 does not use the packet
protocol.

Change-Id: I1e245462b64ec7e6f4c1be9406efe445b30264fa
Origin-Change-Id: I25d4d911bab6a7f23549178bb1ca80db9a0a4d99
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/109048
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>