Adding _POSIX_C_SOURCE to LOCAL_CFLAGS
BZ: 105424

__POSIX_VISIBLE will be set to 198808 if _POSIX_SOURCE is defined and
_POSIX_C_SOURCE is not.  This will lead to some bionic/libm
functions invisible to this module.  Not defining any of them would
lead to __POSIX_VISIBLE being set to a safe, default value.

However, either _POSIX_SOURCE or _POSIX_C_SOURCE need to be defined
when using alsa-lib, or it would lead to redefining timeval and
timespec.

So, we have to use _POSIX_C_SOURCE=200809 (the default value if it
wasn't set)

Change-Id: I191a792c59bea92cf70a12301e751ce7d60138d0
Signed-off-by: Yong Yao <yong.yao@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>
Reviewed-on: http://android.intel.com:8080/106586
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>