[PORT FROM R4.2] HWC: fix scramble bar when playing protected content
BZ: 99843

This is the HWC patch to fix scrambled status bar when playing
protected content. It checks video playing status from MDS and
see if using RGBOverlay. Once video is about to play, it stops
using RGBOverlay. The scramble is due to IED is enabled in
video decoder initialization but overlay is showing normal content.

Change-Id: I49ccd249053611b316683ead47cc7cdf6eaaa0f6
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/102504
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>