[PORT FROM MAIN][AD_PROXY] Ignore baud rate command from AuViD
BZ: 89322

AuViD may send a baud rate command useless for ad proxy which should
not be forwarded to the Audience chip.

This patch updates ad proxy to ignore such command and send it back
to AuViD.

The write operation to the chip is moved in a new WRITE state of the
proxy state machine in order to not write a such command to the chip.

Change-Id: I75c41262c82ba974197eb2e881ae1d4ee120fbcb
Origin-Change-Id: I1bf58849fb87603116280f1d5514a9f39679cd4f
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/109054
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>