HWC: fix scramble bar when playing protected content
BZ: 99843

This is the HWC patch to fix scrambled status bar when playing
protected content. It checks video playing status from MDS and
see if using RGBOverlay. Once video is about to play, it stops
using RGBOverlay. The scramble is due to IED is enabled in
video decoder initialization but overlay is showing normal content.

Change-Id: I808a576e9bbebc3f2c2ead6e4d865d9e0cd4bedc
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/104085
Reviewed-by: Zhu, Tianyang <tianyang.zhu@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>