widi: Move libhwcwidi out of hwc and adopt API revisions
BZ: 103724

The frame server API (libhwcwidi) is going to be implemented by
multiple hardware composers for different platforms, so in order to
have just one copy it is being moved into the widi project. So this
patch deletes it from the CTP hwcomposer.

We've made some changes to the API, so additional changes are made
here to adapt to them. The API now supports configuring scaling via
setResolution() and allows hwcomposer to provide clone mode content to
widi instead of widi getting the frames from SurfaceFlinger.

For now, we're not supporting either scaling or providing clone mode.
This patch just makes hwcomposer compatible with the API revisions.
When scaling support is added, we will be more compliant with
Miracast, and the clone mode support will allow us to improve
efficiency.

Change-Id: I4c7751376b17e952f8e6cea0692a84eeb635a8e0
Signed-off-by: Brian Rogers <brian.e.rogers@intel.com>
Reviewed-on: http://android.intel.com:8080/104848
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Singhi, Ashish <ashish.singhi@intel.com>
Reviewed-by: Zhang, Lei <lei.zhang@intel.com>
Reviewed-by: Crabtree, Robert <robert.crabtree@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>