hwc: adjust sprite pixel format for premult buffer
BZ: 101151

When setting with BGRA, sprite plane still does alpha blending for
premult buffer. So we will get black for area with alpha == 0.

Setting sprite plane to ignore alpha can correct the display. We only do
this if this layer is the bottom layer or the layer doesn't cover any
other layers underneath it.

Change-Id: I36ea78922ee150b5b22d70cec31cc391d6e955e4
Signed-off-by: Hong Liu <hong.liu@intel.com>
Reviewed-on: http://android.intel.com:8080/104878
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>