[PORT FROM R42] hwc: free hdmi device resource after plugout
BZ: 103001

When playing WV video in extended mode, if we plug out HDMI cable,
next round on MIPI device prepare, it tries to find an free overlay
plane to attach the overlay layer. But since the overlay plane is still
attached to HDMI device and HDMI device doesn't have chance to free it.
The video is not shown on local MIPI correctly.

Fixed by freeing all planes and layer list in HDMI display device when
HDMI is plugged out.

Change-Id: I902ceae442c891ee87e6312e955b2cf83d0fc859
Signed-off-by: Hong Liu <hong.liu@intel.com>
Signed-off-by: Randy Xu <randy.xu@intel.com>
Signed-off-by: Tong, Bo <box.tong@intel.com>
Reviewed-on: http://android.intel.com:8080/111343
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Liu, MinX Z <minx.z.liu@intel.com>
Tested-by: Liu, MinX Z <minx.z.liu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>