hwc: free hdmi device resource after plugout
BZ: 103001

When playing WV video in extended mode, if we plug out HDMI cable,
next round on MIPI device prepare, it tries to find an free overlay
plane to attach the overlay layer. But since the overlay plane is still
attached to HDMI device and HDMI device doesn't have chance to free it.
The video is not shown on local MIPI correctly.

Fixed by freeing all planes and layer list in HDMI display device when
HDMI is plugged out.

Change-Id: I0f5c154e02a20a351d10cb618aa99de42eb3a16b
Signed-off-by: Hong Liu <hong.liu@intel.com>