HWC: fix blank screen of RsFountain apk on victoriabay
BZ: 92664

This is the patch to fix black screen of RsFountain apk on VB.
It is due to the format of the plane is set to be BGRA, which
will do blending by using alpha channel with background. The
alpha channel of this apk is zero and thus cause the blending
result is blank.
To fix it, when using framebuffer without YUV layers, we force
to use BGRX format in order to bypass alpha channel.

Change-Id: Idd3c2f7c88f78c3e86936f75b96f224d9abb944f
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/104865
Reviewed-by: Liu, MinX Z <minx.z.liu@intel.com>
Tested-by: Liu, MinX Z <minx.z.liu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>