Fix RGBOverlay fail in landscape mode
BZ: 110393

There are 2 regressions in RGB overlay path.
1. RGBOverlay can't work in landscape mode and it will crash HWC
   sometimes.
2. RGBOverlay will flash green bar when it exit.

The first issue occurs when there is only a StatusBar in landscape mode.
Patch 108045 has changed the NV12 buffer's width to original width
and hwc will calculate the stride itself. But the calculation is
wrong for small width. This will cause overlay update failed.
There is another issue in hwc, we shouldn't use RGBOverlay if there
is only one layer. Also for the revisitLayerList if updateLayersData
failed, we shouldn't attach any new layers to h/w because we will not
update plane context for it.

The second is caused by patch 104865. The patch will change z-order
to make overlay on top if RGB overlay enabled. But it didn't change
z-order back even there are 2 rgb layers since the condition is no
YUV layer. hwc code has another hack to hide overlay plane with 3 prepare
/commit delay(it's urgly, but I don't know why it is still there). The
overlay will flash a green bar in short time because the plane buffer
is freed and overlay plane is still on top.

Change-Id: I6af8d81971ba7e455a473b98675a0c2aee59fe1a
Signed-off-by: Marc Mao <marc.mao@intel.com>
Reviewed-on: http://android.intel.com:8080/110905
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Liu, MinX Z <minx.z.liu@intel.com>
Tested-by: Liu, MinX Z <minx.z.liu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>