widi: Optimize widi presentation mode path
BZ: 118159

Currently BG/presentation mode video playback is streamed at local display resolution
instead of content resolution. This is resulted in unnecessary scaling and color conversions.

This patch optimizes BG mode video playback path and RGB data path if
1. There is only one RGB or NV12 layer
2. there is no transformation and
3. there is no blending

These optimizations are done by setting overlay flag when the above conditions are met.

If the layer is NV12, HWC sends the decoder's khandles to Widi pipeline(same as in extended mode).
If the layer is RGB, we allocate NV12 buffers to blit the RGB framebuffer into, and then
send those buffers to widi pipeline.

This patch is ported from r42tw (patch# 112242) and modified as needed.

Change-Id: I5c36867414f8c3499d1f5a14483222295e9f7d46
Signed-off-by: Brian Rogers <brian.e.rogers@intel.com>
Signed-off-by: mamatha balguri <mamatha.balguri@intel.com>
Reviewed-on: http://android.intel.com:8080/115227
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>