[MERR][Workaround] After resume 5.1 channel file playback slowly and have noise through HDMI 5.1 device
BZ: 116748

In this scenario when the playback is resumed, because of touch
tone there will be a situation, where 2 write calls get invoked,
one for writing the stereo content and other for multi-channel
content which will cause this issue.

With this fix when this situation happens, the current buffer is
ignored and not written to pcm device.

Category: Feature Differentiation
Domain: AUDIO_MEDIA - HDMI Multichannel
Severity: High
Origin: Internal
Upstream-Candidate: No, needs rework

Change-Id: I8bacfb972308d60a79a575e52f73bd9e42a2850d
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>