Audio does not sync well at HDMI hotplug when playing a clip
BZ: 139820

When he hdmi hotplug event is delayed to the audio service,
the buffers written to the device are not stopped till the
event is recieved even though there is a failure in the writes.
So the same number of bytes are returned to the player and
this is used in progress bar updates. Since the data is not
rendered, the progress bar updates will be fast due to
immediate return on write failure.

Category : Device enablement
Domain   : AUDIO_MEDIA-HDMI
Origin   : Internal
Upstream-Candidate : no, not aosp

Change-Id: I4b1c0c013dc122018fa49032dfabd1e8e7a3ed84
Signed-off-by: jerome <jerome.anand@intel.com>
Reviewed-on: http://android.intel.com:8080/134468
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Reviewed-by: Kandasamy, Muthukumar <muthukumar.kandasamy@intel.com>
Reviewed-by: Krishna, Samaga <samaga.krishna@intel.com>
Reviewed-by: Martin, Antoine <antoine.martin@intel.com>
Tested-by: Martin, Antoine <antoine.martin@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>