hwc: fix hwc crash when playing potected content in WV over HDMI
BZ: 157692

This crash is because a NULL buffer handle passed to gralloc and
cause gralloc crash. The NULL buffer handle is because overlay is
disabled from MDS at the begining of video playing, but as it is
protected video, the compositon type is still set to be HWC_OVERLAY.

This patch fix the crash by checking overlay status in overlay flip.
Meanwhile, add handle check after plane flip to make sure no NULL
handle assigned to gralloc.

Change-Id: I2271eaa254c91440df1c10d5710f62c4e987dd4a
Signed-off-by: Lei Zhang <lei.zhang@intel.com>